requirements.txt
Flask
gunicorn
ultralytics
opencv-python-headless
numpy
Pillow
cloudinary
python-dotenv
reportlab
